Move utils into package, create factories for object creation, setup basics for social feed algorithm (unit tested).
